Refactor: - Разделены тесты UserService на отдельные классы: создание, обновление, удаление, получение, валидация - Введён базовый класс AbstractUserServiceTest для повторного использования общих настроек - Все тесты проходят успешно, покрытие улучшено
