+++ homework_4/src/test/java/ru/aston/teamwork/AbstractIntegrationTest.java
@@ -1,14 +1,11 @@
 package ru.aston.teamwork;
 
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
-import org.springframework.transaction.annotation.Transactional;
 import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
@@ -18,7 +15,7 @@ import ru.aston.teamwork.repository.UserRepository;
 @SpringBootTest
 @Testcontainers
 @ActiveProfiles("test")
-public class AbstractIntegrationTest {
+public abstract class AbstractIntegrationTest {
 
     @Autowired
     private UserRepository userRepository;
@@ -40,7 +37,6 @@ public class AbstractIntegrationTest {
         registry.add("spring.datasource.password", postgres::getPassword);
     }
 
-    @Transactional
     @BeforeEach
     void clearDatabase() {
         userRepository.deleteAll();
